# Package updates & TS startup time optimizations