--- a/metaspace/graphql/package.json
+++ b/metaspace/graphql/package.json
@@ -92,7 +92,7 @@
     "reflect-metadata": "^0.1.10",
     "sprintf-js": "^1.0.3",
     "subscriptions-transport-ws": "^0.9.16",
-    "ts-node": "^8.3.0",
+    "ts-node": "^10.0.0",
     "typeorm": "0.2.25",
     "typescript": "^3.9.2",
     "uuid": "^3.3.2",
@@ -116,7 +116,7 @@
     "request-promise": "^4.2.1",
     "supertest": "^4.0.2",
     "ts-jest": "25.5.0",
-    "ts-node-dev": "^1.0.0-pre.51"
+    "ts-node-dev": "^1.1.8"
   },
   "jest": {
     "roots": [