--- a/metaspace/webapp/package.json
+++ b/metaspace/webapp/package.json
@@ -102,7 +102,7 @@
     "babel-loader": "^8.2.2",
     "babel-plugin-component": "^1.1.1",
     "babel-preset-vca-jsx": "^0.3.4",
-    "compression-webpack-plugin": "^8.0.1",
+    "compression-webpack-plugin": "^6.1.1",
     "cross-env": "^3.0.0",
     "eslint": "^6.8.0",
     "eslint-plugin-import": "^2.20.0",