--- a/metaspace/webapp/package.json
+++ b/metaspace/webapp/package.json
@@ -123,7 +123,7 @@
     "vue-cli-plugin-apollo": "^0.21.3",
     "vue-cli-plugin-svg": "^0.1.3",
     "vue-template-compiler": "2.6.11",
-    "webpack": "^5.42.0",
+    "webpack": "^4.46.0",
     "webpack-bundle-analyzer": "^3.6.0",
     "webpack-cli": "^4.7.2"
   },